cmd/flash_cards/main_test.go
package main

import (
	"bufio"
	"fmt"
	"strings"
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestPlay(t *testing.T) {
	t.Log("https://hyperskill.org/projects/224")

	tests := []struct {
		give string
		want string
	}{
		{
			`2
print()
outputs text
str()
converts to a string
outputs text
outputs text`,
			`Input the number of cards:
The term for card #1:
The definition for card #1:
The term for card #2:
The definition for card #2:
Print the definition of "print()":
Correct!
Print the definition of "str()":
Wrong. The right answer is "converts to a string".
`,
		},
	}

	for i, tt := range tests {
		t.Run(fmt.Sprintf("%d", i), func(t *testing.T) {
			scanner := bufio.NewScanner(strings.NewReader(tt.give))
			buf := new(strings.Builder)
			writer := bufio.NewWriter(buf)

			FlashCards(scanner, writer)

			err := writer.Flush()
			assert.NoError(t, err)

			got := buf.String()
			assert.Equal(t, tt.want, got)
		})
	}
}
